requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain-community==0.0.29
python-dotenv
jupyterlab
ipykernel
